+++ Game/road.h
@@ -12,7 +12,7 @@ class road: public entity
 public:
     road(float r_w, float r_h, SDL_Texture* r_tex);
     void textureRoad(renderWindow& a);
-    void animateRoad(renderWindow& a, int speed);
+    void animateRoad(renderWindow& a, float speed);
 private:
     std::vector <SDL_Texture*> tex;
     std::vector <entity> frame;
